Fixes mismatching of remote names when task sent to All-Pairs multicore
