+++ dttools/src/text_list.h
@@ -7,6 +7,5 @@ char *text_list_get(struct text_list *t, int i);
 int text_list_append(struct text_list *t, const char *item);
 int text_list_size(struct text_list *t);
 void text_list_delete(struct text_list *t);
-struct text_list *text_list_allpairs_remote_create(const char *path, const char *set);
 
 #endif
